requirements.txt
-f ./prebuilt

Django==4.1
djangorestframework
bosdyn-client >= 3.1
bosdyn-mission
bosdyn-choreography-client
Pillow==6.1.0
six
pyqt5
python-dotenv
numpy
opencv-python